+++ internal/data/parsers/copy.go
