+++ internal/data/parsers/fkey.go
